Cambios para mejorar la vista

+++ operationslibrary/src/main/java/com/elizabeth/operationslibrary/Operationes.kt
@@ -3,18 +3,21 @@ package com.elizabeth.operationslibrary
 class Operations {
 
     //Funcion Suma
-    fun sum(a: Int, b: Int): Int {
+    fun sum(a: Double, b: Double): Double {
         return a + b
 
     }
-    fun subtract(a: Int, b: Int): Int {
+
+    fun subtract(a: Double, b: Double): Double {
         return a - b
 
     }
-    fun multiply(a: Int, b: Int): Int {
+
+    fun multiply(a: Double, b: Double): Double {
         return a * b
 
     }
+
     fun divide(a: Double, b: Double): Double {
         return a / b
